[Engine3D] Fix scroll container not smooth issue
Why: 1. Fix float scroller issue cause by precision.
     2. Apply slop state instead of min velocity in fling judgment.
How: 1. Float precision casting on distance formula.
     2. Apply slop state instead of min velocity in fling judgment.
Verify steps: N/A
Possible effects: none

+++ src/min3d/FloatOverScroller.java
@@ -572,7 +572,7 @@ public class FloatOverScroller {
         private int mSplineDuration;
 
         // Distance to travel along spline animation
-        private int mSplineDistance;
+        private float mSplineDistance;
 
         // Whether the animation is currently in progress
         private boolean mFinished;
@@ -772,7 +772,7 @@ public class FloatOverScroller {
                 totalDistance = getSplineFlingDistance(velocity);
             }
 
-            mSplineDistance = (int) (totalDistance * Math.signum(velocity));
+            mSplineDistance = (float) (totalDistance * Math.signum(velocity));
             mFinal = start + mSplineDistance;
 
             // Clamp to a valid final position
